+++ MK3166/MK3166-1.0.0/cores/arduino/RingBuffer.h
@@ -38,13 +38,13 @@ public:
      */
     int puts(uint8_t *data, int len);
 
-    uint8_t peek();
+    int peek();
     
     /** get from ring buffer
      * @param dat data
      * @retval 0:ok / -1:error
      */
-    uint8_t getc();
+    int getc();
 
     /** get from ring buffer
      * @param dat data
